+++ src/main/scala/com/particeep/api/models/fundraise/loan/FundraiseLoanEdition.scala
@@ -15,7 +15,7 @@ case class FundraiseLoanEdition(
   recipient_type:      Option[String]           = None,
   start_at:            Option[ZonedDateTime]    = None,
   end_at:              Option[ZonedDateTime]    = None,
-  amount_target:       Option[Int]              = None,
+  amount_target:       Option[Long]             = None,
   currency:            Option[Currency]         = None,
   score:               Option[String]           = None,
   tags:                Option[String]           = None,
