+++ src/main/scala/com/particeep/api/models/fundraise/reward/FundraiseRewardCreation.scala
@@ -19,7 +19,7 @@ case class FundraiseRewardCreation(
   recipient_type:      Option[String],
   start_at:            Option[ZonedDateTime] = None,
   end_at:              Option[ZonedDateTime] = None,
-  amount_target:       Int,
+  amount_target:       Long,
   currency:            Option[Currency]      = None,
   tags:                Option[String]        = None
 )
